refactor(hosts): separate availability routing from recovery ownership
