+++ CONTEXT.md
@@ -237,6 +237,17 @@ _Avoid_: Scoped search (Scope is the ownership primitive), treating `@` as a six
 `src/core/services/hosts` owns machine lifecycle and workspace-server provisioning under the Host vocabulary (`HostService`, the `hosts` wire domain) — the merged home of the former `remote-machine` and `workspace-server` services. "Machine" remains a UI label only. Host (wire) remains the separately-scoped wire term.
 _Avoid_: Machine (outside UI labels), "remote" as a noun
 
+**Host service**:
+The desktop's services for one remote Host identity: connection intent, runtime access, and workspace
+server maintenance. `Hosts` owns lookup and replacement; retained HostService instances are invalid
+after machine identity edits. A registry lease follows replacements until its owner is disposed.
+_Avoid_: Confusing the desktop `Hosts` service with the Host-authoritative workspace registry
+
+**Host workspace server control**:
+The desktop's installation and daemon-control service for one remote Host (`host.server`). It owns
+maintenance sequencing and daemon state observation. Connection recovery remains the supervisor's job.
+_Avoid_: Equating Disconnect with stopping a daemon shared by multiple desktop clients
+
 **Host availability**:
 The Hosts-domain fact describing whether a Host runtime can currently serve desktops. SSH connected
 only begins preparation; availability becomes ready after the runtime handshake. The Hosts domain
