+++ apps/emdash-desktop/src/core/services/hosts/api/availability.ts
@@ -1,9 +1,10 @@
 import { hostRefSchema, type HostRef } from '@emdash/core/primitives/host/api';
 import {
   runtimeResolveErrorSchema,
+  runtimeHostUnavailable,
   type RuntimeResolveError,
 } from '@emdash/core/primitives/runtime-resolution/api';
-import type { Result } from '@emdash/shared';
+import { err, ok, type Result } from '@emdash/shared';
 import type { Scope } from '@emdash/shared/concurrency';
 import type { Readable } from '@emdash/wire/state';
 import { z } from 'zod';
@@ -40,26 +41,37 @@ export type HostReady = z.output<typeof hostReadySchema>;
 export type HostPreparingPhase = Extract<HostAvailabilityState, { kind: 'preparing' }>['phase'];
 export type RecoveryCause = 'demand' | 'connect' | 'retry' | 'ssh-edge' | 'online' | 'focus';
 export type ExplicitRecoveryCause = Extract<RecoveryCause, 'connect' | 'retry'>;
-export type HostDemandMode = 'automatic' | 'passive';
 export type HostWakeCause = Extract<RecoveryCause, 'ssh-edge' | 'online' | 'focus'>;
 export type BrowserHostWakeCause = Extract<HostWakeCause, 'online' | 'focus'>;
-
-export interface HostDemandLease {
-  readonly mode: HostDemandMode;
-  setMode(mode: HostDemandMode): void;
-}
 
 export interface HostAvailability {
   state(host: HostRef): Readable<HostAvailabilityState>;
   stateFor(host: HostRef): HostAvailabilityState;
   requireReady(host: HostRef): Result<HostReady, RuntimeResolveError>;
-  demand(host: HostRef, mode: HostDemandMode, owner: Scope): HostDemandLease;
+  lease(host: HostRef, owner: Scope): void;
   wake(host: HostRef, cause: HostWakeCause): void;
   wakeDemanded(cause: BrowserHostWakeCause): void;
   ensureReady(host: HostRef, cause: RecoveryCause): Promise<Result<HostReady, RuntimeResolveError>>;
   requestReady(host: HostRef, cause: ExplicitRecoveryCause): void;
   invalidate(host: HostRef, issue?: RuntimeResolveError): void;
   suspend(host: HostRef): void;
+}
+
+/** Converts observed availability to an immediate result; never starts connection work. */
+export function hostReadyResult(
+  host: HostRef,
+  state: HostAvailabilityState
+): Result<HostReady, RuntimeResolveError> {
+  if (state.kind === 'ready') return ok({ host, generation: state.generation });
+  if (state.kind === 'unavailable' && state.issue) return err(state.issue);
+  if (state.kind === 'preparing') {
+    return err(
+      state.phase === 'connecting'
+        ? runtimeHostUnavailable(host, 'connection-failed', 'Host connection is not ready')
+        : runtimeHostUnavailable(host, 'runtime-unavailable', 'Host runtime is not ready')
+    );
+  }
+  return err(runtimeHostUnavailable(host, 'offline', 'Host is offline'));
 }
 
 export function runtimeRecoveryDisposition(
