+++ apps/emdash-desktop/src/core/services/hosts/node/index.ts
@@ -2,7 +2,5 @@ export {
   createHostAvailability,
   HostAvailabilityService,
   type CreateHostAvailabilityOptions,
-  type HostReadinessAdapter,
-  type HostReadinessContext,
 } from './availability';
 export { translateHostPreparationError } from './runtime-resolution';
