+++ apps/emdash-desktop/src/core/services/hosts/node/wire-controller.test.ts
@@ -9,16 +9,37 @@ import { hostsContract } from '../api';
 import { createHostAvailability } from './availability';
 import type { Hosts } from './hosts';
 import { createHostsWireController } from './wire-controller';
+import { createWorkerHostAvailability } from './worker-host-availability';
 
 describe('Hosts Wire availability', () => {
   it('publishes readiness through the Host-keyed live state', async () => {
     const scope = createScope({ label: 'hosts-wire-availability-test' });
-    const availability = createHostAvailability({
+    const workers = createWorkerHostAvailability({
       scope,
       readiness: { prepare: async () => ok() },
     });
-    const wakeDemanded = vi.spyOn(availability, 'wakeDemanded');
     const host = hostRef('remote', 'ssh-1');
+    const availability = createHostAvailability({
+      scope,
+      local: workers,
+      remote: () => ({
+        connection: {
+          availability: workers.state(host),
+          lease: (owner) => workers.lease(host, owner),
+          pin: async () => ok(),
+          disconnect: async () => {
+            workers.suspend(host);
+            return ok();
+          },
+        },
+        waitUntilReady: () => workers.ensureReady(host, 'demand'),
+      }),
+      remoteState: () => workers.state(host),
+      remoteLease: (_id, owner) => workers.lease(host, owner),
+      wakeRemote: (cause) => workers.wakeDemanded(cause),
+      revalidateRemote: () => workers.invalidate(host),
+    });
+    const wakeDemanded = vi.spyOn(availability, 'wakeDemanded');
     const serverStates = expose(hostsContract.serverStates, { runtime: cell({}) });
     const service = { stateModel: { host: serverStates } } as Hosts;
     const disconnect = vi.fn(async () => {
